Only handle the size of queue_size inside DeviceNode, and use unsigned type externally

+++ include/uorb/publication.h
@@ -45,7 +45,7 @@ namespace uorb {
 /**
  * uORB publication wrapper class
  */
-template <const orb_metadata &meta, uint8_t queue_size = 1>
+template <const orb_metadata &meta, uint16_t queue_size = 1>
 class Publication {
   using Type = typename msg::TypeMap<meta>::type;
 
@@ -71,7 +71,7 @@ class Publication {
 /**
  * The publication class with data embedded.
  */
-template <const orb_metadata &T, uint8_t queue_size = 1>
+template <const orb_metadata &T, uint16_t queue_size = 1>
 class PublicationData : public Publication<T, queue_size> {
   using Type = typename msg::TypeMap<T>::type;
